+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to TTT2 will be documented here. Inspired by [keep a changel
 - Added the beacon back into TTT2, an equipment that was disabled long ago in base TTT
   - Can only be bought by policing roles
   - Creates a wallhack in a sphere around it, which is visible to everyone
+- Added new hooks `TTT2BeaconDetectPlayer` and `TTT2BeaconDeathNotify` to allow preventing / overriding a beacon's player detection & alerts (by @spanospy)
 
 ### Changed